Main_Project/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
django-cors-headers>=4.2.0

